moved id generation function to utilities and implemented id generation for tags
