+++ src/api/lambdas/bedrock-api-backend/assets/getTasks.js
@@ -1,9 +1,7 @@
-/* eslint-disable import/extensions */
 /* eslint-disable no-console */
 import pgpkg from 'pg';
-import pgErrorCodes from '../pgErrorCodes.js';
-
 const { Client } = pgpkg;
+import pgErrorCodes from '../pgErrorCodes.js';
 
 async function newClient(connection) {
   const client = new Client(connection);
@@ -27,7 +25,7 @@ async function readTasks(client, assetName) {
 }
 
 function formatTasks(res) {
-  const tempTasks = [];
+  let tempTasks = [];
   for (let i = 0; i < res.rowCount; i += 1) {
     tempTasks.push(
       {
@@ -69,7 +67,7 @@ async function getTasks(pathElements, queryParams, connection) {
     if (res.rowCount === 0) {
       response.message = 'No tasks found';
     } else {
-      tasks = formatTasks(res);
+      tasks = formatTasks(res)
       response.result = {
         items: tasks,
       };
@@ -80,9 +78,9 @@ async function getTasks(pathElements, queryParams, connection) {
     response.message = error.message;
     return response;
   } finally {
-    await client.end();
+    await client.end()
+    return response;
   }
-  return response;
 }
 
 export default getTasks;
